添加时钟功能

+++ navy-apps/libs/libndl/include/NDL.h
@@ -9,7 +9,7 @@ extern "C" {
 
 int NDL_Init(uint32_t flags);
 void NDL_Quit();
-uint32_t NDL_GetTicks();
+uint64_t NDL_GetTicks();
 void NDL_OpenCanvas(int *w, int *h);
 int NDL_PollEvent(char *buf, int len);
 void NDL_DrawRect(uint32_t *pixels, int x, int y, int w, int h);
